+++ src/Components/Core/Application/DataTransferObjects/LearningSpacePrecursorAndSuccessorTO.ts
@@ -3,4 +3,5 @@ import LearningSpaceTO from "./LearningSpaceTO";
 export default class LearningSpacePrecursorAndSuccessorTO {
   precursorSpaces: LearningSpaceTO[];
   successorSpaces: LearningSpaceTO[];
+  availableSpaces: LearningSpaceTO[];
 }